# Conflicts:
#	.github/workflows/run-tests-linux.yml
#	build.sbt
#	javalib/src/main/scala/java/io/File.scala
#	javalib/src/main/scala/java/lang/StackTraceElement.scala
#	javalib/src/main/scala/java/lang/Throwables.scala
#	javalib/src/main/scala/java/lang/impl/PosixThread.scala
#	javalib/src/main/scala/java/net/AbstractPlainSocketImpl.scala
#	javalib/src/main/scala/java/net/SocketHelpers.scala
#	javalib/src/main/scala/java/nio/channels/FileChannelImpl.scala
#	javalib/src/main/scala/java/nio/channels/FileLockImpl.scala
#	javalib/src/main/scala/java/util/Date.scala
#	nativelib/src/main/scala/scala/scalanative/runtime/Intrinsics.scala
#	nativelib/src/main/scala/scala/scalanative/unsafe/CVarArgList.scala
#	nativelib/src/main/scala/scala/scalanative/unsafe/package.scala
#	nir/src/main/scala/scala/scalanative/nir/Versions.scala
#	nscplugin/src/main/scala-2/scala/scalanative/nscplugin/NirPrimitives.scala
#	test-interface/src/main/scala/scala/scalanative/testinterface/signalhandling/SignalConfig.scala
#	unit-tests/native/src/test/scala/java/lang/SystemWithPosixTest.scala
#	unit-tests/native/src/test/scala/java/net/UdpSocketTest.scala
#	unit-tests/native/src/test/scala/scala/scalanative/posix/TimeTest.scala
#	unit-tests/native/src/test/scala/scala/scalanative/unsafe/CVarArgListTest.scala
#	unit-tests/native/src/test/scala/scala/scalanative/unsafe/TagTest.scala
#	unit-tests/shared/src/test/scala/javalib/lang/ThrowablesTest.scala

+++ javalib/src/main/scala/java/math/BigDecimal.scala
@@ -1024,7 +1024,7 @@ class BigDecimal() extends Number with Comparable[BigDecimal] {
             if (compRemDiv0 == 0) {
               val bi = qr.rem.multiply(powerOf10(exp2))
               val rem = bi.divide(divisor.getUnscaledValue())
-              Math.abs(rem.signum())
+              Math.abs(rem.signum()).toLong
             } else {
               compRemDiv0
             }
